requirements.txt
hydra-core==1.2.0
torch==1.12.0
torchvision==0.13.0
matplotlib==3.3.4
scikit-learn
seaborn
tqdm
gensim
ipykernel
protobuf
regex
slurmpy
wordsegment
prettytable
argunparse
tensorboard
transformers
nltk
opencv-python
einops
tensorboard
spacy
sentence-transformers
git+https://github.com/openai/CLIP.git
tabulate